+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## IN DEVELOPMENT
+
 ## v0.5.0 - [October 6, 2021](https://github.com/lando/cli/releases/tag/v0.5.0)
 
 Lando is **free** and **open source** software that relies on contributions from developers like you! If you like Lando then help us spend more time making, updating and supporting it by [contributing](https://github.com/sponsors/lando).
 
 * First release of `platformsh` as an external plugin!
+* Added testing for most `services` [#3]
+* Fixed some bugs